Import live.2013.09.18.tar.gz
2013.09.11
- It turns out that the fix that we'd made back in version 2012.10.04 to (try to) prevent RTCP reports from being sent over
  TCP connections prematurely was incomplete.  (It had prevented only "SR" reports - not "RR" - from being sent.)
  This version fixes this.  (Thanks to Stanley Biggs for reporting this.)
- Made an update to the previous change to the way that "openRTSP" handles the "-Q" option.

2013.09.18
- Fixed an obscure bug - triggered by the "ProxyServerMediaSession" code - that could sometimes cause a previously deleted object
  to be accessed.  (Thanks to Yogev Cohen for reporting this.)

+++ testProgs/playCommon.cpp
@@ -1046,7 +1046,7 @@ static void scheduleNextQOSMeasurement() {
   struct timeval timeNow;
   gettimeofday(&timeNow, NULL);
   unsigned timeNowUSecs = timeNow.tv_sec*1000000 + timeNow.tv_usec;
-  int64_t usecsToDelay = nextQOSMeasurementUSecs - timeNowUSecs;
+  int usecsToDelay = nextQOSMeasurementUSecs - timeNowUSecs;
 
   qosMeasurementTimerTask = env->taskScheduler().scheduleDelayedTask(
      usecsToDelay, (TaskFunc*)periodicQOSMeasurement, (void*)NULL);
